Fix texture cell sizing
The sprite sheet cell size was being calculated using integer division because m_cols and m_rows were integers. This change casts them to float so texture frame dimensions are computed correctly. It shouldnt have caused errors due to the fact that I want the cells to be whole number sizes, but it was giving a compile time error so I changed it.

+++ GameEngine/Source/Engine/Renderer/TexFrames.cpp
@@ -35,7 +35,7 @@ namespace sr {
 
         Vector2 textureSize = m_texture->GetSize();
 
-        m_cellSize = textureSize / Vector2{ m_cols, m_rows };
+        m_cellSize = textureSize / Vector2{ (float)m_cols, (float)m_rows };
 
         return true;
     }
